#82 bubbled out TableColumnsIWithIndex
